fix: FileStatus and move directory
- FileStatusをシンプルに実装
- ディレクトリを移動
