+++ 07.ls_object/lib/file/file_data.rb
